+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 
 - Support for namespace imports in TypeScript/JavaScript (#190)
+- Added a configurable resource cutoff for policy generation resource lists.
 
 ### Fixed